LPD-79446 Rename date-filter directory to snake_case
