+++ modules/apps/export-import/export-import-web/src/main/resources/META-INF/resources/revamp/js/pages/export/NewExport.tsx
@@ -9,7 +9,7 @@ import {FormikValues} from 'formik';
 import React from 'react';
 
 import {Wizard, WizardStep} from '../../components/Wizard';
-import {DateFilterValues} from '../../components/date-filter';
+import {DateFilterValues} from '../../components/date_filter';
 import DataSelectionStep from './steps/DataSelectionStep';
 import SettingsStep from './steps/SettingsStep';
 import SetupStep from './steps/SetupStep';
